Refactor code to work with >>= syntax
I've recombined the previous function which worked to main, so that I
could understand it myself. The types were previously working but I had
no real understanding of what was happening. Not having a grasp on do
syntax made it worse, as I did not understand why the types were
failing.

In following commits, I will tease apart the functions again